+++ nginx.conf
@@ -1,16 +1,13 @@
-
 server {
     listen 80;
     server_name localhost;
     root /usr/share/nginx/html;
     index index.html;
 
-    # Handle SPA routing and prevent caching of main index
+    # Handle SPA routing
     location / {
         try_files $uri $uri/ /index.html;
-        expires -1;
         add_header Cache-Control "no-store, no-cache, must-revalidate";
-        add_header Pragma "no-cache";
     }
 
     # Specific no-cache for env-config.js
@@ -24,11 +21,5 @@ server {
     location /assets {
         expires 1y;
         add_header Cache-Control "public, no-transform";
-        try_files $uri $uri/ =404;
     }
-
-    # Additional security headers
-    add_header X-Frame-Options "SAMEORIGIN";
-    add_header X-Content-Type-Options "nosniff";
-    add_header X-XSS-Protection "1; mode=block";
 }
