+++ package.json
@@ -57,7 +57,6 @@
     "markdown-it-highlightjs": "^4.2.0",
     "markdown-it-katex": "^2.0.3",
     "next-themes": "^0.3.0",
-    "octokit": "^4.1.0",
     "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18.3.1",
