+++ public/env-config.js
@@ -3,7 +3,6 @@ window.env = {
   VITE_N8N_WEBHOOK_URL: "",
   VITE_WELCOME_MESSAGE: "",
   VITE_SITE_TITLE: "",
-  VITE_N8N_WEBHOOK_USERNAME: "",
-  VITE_N8N_WEBHOOK_SECRET: "",
-  VITE_ASSISTANT_NAME: "Talkflow"
+  VITE_ASSISTANT_NAME: ""
 };
+
